refactor: add unit-communitydetail module
