+++ core-commonui/detailopener-impl/build.gradle.kts
@@ -39,6 +39,7 @@ kotlin {
                 implementation(projects.coreNavigation)
                 implementation(projects.coreCommonui)
                 implementation(projects.coreCommonui.unitPostdetail)
+                implementation(projects.coreCommonui.unitCommunitydetail)
                 implementation(projects.coreCommonui.detailopenerApi)
 
                 implementation(projects.domainLemmy.data)
