+++ core-commonui/src/androidMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/di/Utils.kt
@@ -1,22 +1,10 @@
 package com.github.diegoberaldin.raccoonforlemmy.core.commonui.di
 
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.communitydetail.CommunityDetailMviModel
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.userdetail.UserDetailMviModel
-import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.CommunityModel
 import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.UserModel
 import org.koin.core.parameter.parametersOf
 import org.koin.java.KoinJavaComponent.inject
 
-actual fun getCommunityDetailViewModel(
-    community: CommunityModel,
-    otherInstance: String,
-): CommunityDetailMviModel {
-    val res: CommunityDetailMviModel by inject(
-        clazz = CommunityDetailMviModel::class.java,
-        parameters = { parametersOf(community, otherInstance) },
-    )
-    return res
-}
 
 actual fun getUserDetailViewModel(user: UserModel, otherInstance: String): UserDetailMviModel {
     val res: UserDetailMviModel by inject(
