+++ core-commonui/src/iosMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/di/Utils.kt
@@ -1,32 +1,15 @@
 package com.github.diegoberaldin.raccoonforlemmy.core.commonui.di
 
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.communitydetail.CommunityDetailMviModel
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.userdetail.UserDetailMviModel
-import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.CommunityModel
 import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.UserModel
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import org.koin.core.parameter.parametersOf
 
-actual fun getCommunityDetailViewModel(
-    community: CommunityModel,
-    otherInstance: String,
-): CommunityDetailMviModel =
-    CommonUiViewModelHelper.getCommunityDetailModel(community, otherInstance)
-
 actual fun getUserDetailViewModel(user: UserModel, otherInstance: String): UserDetailMviModel =
     CommonUiViewModelHelper.getUserDetailModel(user, otherInstance)
 
 object CommonUiViewModelHelper : KoinComponent {
-    fun getCommunityDetailModel(
-        community: CommunityModel,
-        otherInstance: String,
-    ): CommunityDetailMviModel {
-        val model: CommunityDetailMviModel by inject(
-            parameters = { parametersOf(community, otherInstance) },
-        )
-        return model
-    }
 
     fun getUserDetailModel(user: UserModel, otherInstance: String): UserDetailMviModel {
         val model: UserDetailMviModel by inject(
