+++ core-commonui/unit-communitydetail/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/communitydetail/di/Utils.kt
@@ -1,0 +1,9 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.communitydetail.di
+
+import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.CommunityModel
+import com.github.diegoberaldin.raccoonforlemmy.unit.communitydetail.CommunityDetailMviModel
+
+expect fun getCommunityDetailViewModel(
+    community: CommunityModel,
+    otherInstance: String = "",
+): CommunityDetailMviModel
