+++ settings.gradle.kts
@@ -59,3 +59,4 @@ include(":core-commonui:unit-createpost")
 include(":core-commonui:unit-postdetail")
 include(":core-commonui:detailopener-api")
 include(":core-commonui:detailopener-impl")
+include(":core-commonui:unit-communitydetail")
